feat(.bashrc): add conda init and base venv activation
Conda initializes your shell with the "conda init" command.
This is adding the respective entry to your .bashrc.
Unfortunately it uses for this static paths to your home dir.
This is not reusable for other machines, if your user is differently
named there.

After replacing it with "$HOME", I found that this section uses single
quotes for the reference to the conda executable. Therefore the variable
$HOME is not resolved. Replaced them therefore also with double quotes.