feat(.vimrc): add undo feature with respective cache
Taken from https://vi.stackexchange.com/a/53
Apparently vim supports after 7.3 an persistent undo cache.
Beforehand everytime one saved and closed a file,
it was not possible to undo changes after re-opening from before.
With this feature, this is now possible.

The answer mentions a cleanup cronjob.
Could be interesting to add this as well to this dotfile repo
and connect it somehow to this commit?