feat(.vimrc): add Vundle as plugin manager
Vundle src: https://github.com/VundleVim/Vundle.vim
Vundle states in its README, that VI-Support should get deactivated.
Besides that it seems to be required that filetype detection needs to be deactivated during Plugin loading.
I activated it afterwards again.

It is stated on https://vim-jp.org/vimdoc-en/filetype.html, that any flags from before the deactivation should be untouched by it.
Therefore the general reactivation via filetype on.

I also used Vundle to install itself and afterwards refer to itself as installed plugin.
This allows to only keep track of the plugins in the dotfiles and not vundle itself explicitly.
If one wants to upgrade vundle, it should be done via vundle itself.
Beforehand, one could have just synced with the master of the Vundle repo again.
There does not seem to be a lot of activity right now. (Last commit August 2019).
So I do not expect the need for updates anyhow.