feat(.vimrc): add file specific folding and comment toggling
add for now support for vim-scripts and python files.
The python comment toggling and folding is picked per default.
Right now the python augroup is useless.
At first I wanted to solve this by using dedicated augroups for vim-scripts, python and a default group.
However I misunderstood how augroups work. I thought, that they overwrite each other's autocommands.
But that is not the case. They just overwrite the autocommands from the same group.
(Given ofc the `autocmd!` in the respective groups.)

Since I like the default python behavior, this should be good for now.
However it can ofc get specialized more in the future.

Source for understanding augroups a little better: https://vi.stackexchange.com/questions/9455/why-should-i-use-augroup

Note: The idea for this generic group is actually something, that I talked to ChatGPT about.
It recommended me it without explaining, that it actually does not work.