feat(.bashrc): add colored git branch into PS1
I found that the current .bashrc still holds the default settings from the regular debian skeleton.
So I invested some time in understanding how the colors in different shells work and how the PS1 is getting interpreted.
There are for example multiple ways on how to represent colors in the terminal.
One way is to use the `tput` tool, which allows for a lot formatting options and a readable usage.
The downside is, that it is not present on every system,
so basing the color settings of this generic .bashrc would be problematic.
Another option is to use ANSI escape codes (https://en.wikipedia.org/wiki/ANSI_escape_code).
There are multiple ways on how to configure this. In the end, it is about the amount of bits getting used for encoding.
Here again the same argument applies - using a broader,
more advanced setting is not necessarily supported on all systems.
Therefore I decided to rely on 8-bits, which ends up with these horrible ESC char encoding (\033).
The 256-bit version for example uses \e instead, which looks much nicer but again, not supported everywhere...

> Why I used a combination of single and double quotes for PS1 content...

Finally I was fighting a little bit with the PS1-Env Variable to control the prompt design.
The goal is to add another field in orange, which represent the current branch of a git project,
that I would be located in at that moment.
An easy solution would be to write an own function that retrieves this branch.
I decided however to rely on the already existing method __git_ps1, which comes with git autocompletion utils.
If those are not present, this is also not shown.
I am however anyway a big fan of autocompletion, so if this is not there,
the system might either have a very short lifecycle, or is not yet fully set up.
When configuring the PS1-Variable, I predefined variables for the color codings.
I found then, that if I use single quotes for the respective string,
the variable are getting interpreted when prompting PS1.
This again seems to prevent the interpretation of the content as color coding
and results into the codes getting literally printed.
When using double quotes however, the color coding works, but the result from calling "__git_ps1" is getting literally
injected into the string (which is in case of the $HOME-dir ordinary an empty string).
So it seems that PS1 needs to get the literal function call instead but the already resolved color codings.
Therefore I relied an a combination of single and double quotes.