fix(.vimrc): prevent indention removal when uncommenting
The provided vim function on <C-k> work perfectly, when no indention is there.
If there is some, the commenting works as well, but uncommenting removes also the whole indention.

I fixed this by using a positive lookbehind for uncommenting.
I also exchanged the commenting function with a search-and-replace approach and hope for no issues,
 since it was before based on using the normal mode.

Overall this is just a trial and error thingy, since I am not quite sure,
 how many side effects can occur due to the auto indention support of different file types.
My hope however is, that the plain search-and-replace approach circumvents these issues.