feat(.vimrc): add shortcut for opening a shell
I found myself recently when developing in vim having the for opening
shells during `pytest` calls. My first solution was using ":!pytest %".
However during the process I needed to see the code in my current buffer
and the results of the tests at the same time. Unfortunately by
switching back to the buffer I was not able to access the test results
anymore. So I now added this shortcut to open a shell for multiple
purposes, which also leaves the shell open after the execution of
arbitrary commands. I am sure that there are more sophisticated
solutions for my original problem, but I think this should in general
increase my motivation to "stay" in vim while developing.

It took me actually some time to come up with a command to ensure that
the shell is always opened on the right side of my screen (currently
there is a lof space there). It seems like one needs to turn vim into
"vertical mode" first and then also enforce here the terminal to be
opened on the right side. I might be still missing points out on the
general handling of windows in vim. My current idea is that the values
of "splitright" and "splitbelow" control if a new window is either
opened left or right or respectively top or bottom.