fix(.bashrc): PS1 escapes for bash line length detection
I realized recently when using my terminal on MacOS with a limit of 80
chars, entering a long command created a line wrap in the current line.
Ordinary it should start a new line for the wrapping. After some
research I found that this is happening due to my $PS1-design.
See also https://askubuntu.com/questions/111840/ps1-problem-messing-up-cli

Therefore I applied the proper escaping directly inside the $PS1 and
moved also the coloring from the git function directly into the setter
of $PS1.
This seems to fix the issue