feat(.vimrc): explicitly set backspace behavior in insert mode
When running vim on my Mac recently I was unable to delete newlines and
indents. On all my linux setups, this was never a problem. Given this
information.
https://vimdoc.sourceforge.net/htmldoc/options.html#%27backspace%27,
it seems like my MacOS setup uses here a different default setting.
Therefore I decided to explicitly add it to my ~/.vimrc