chore: add installer and hint for vim's helptags
While familiarizing with ALE, they claimed a great documentation, using
`:help ale-options`. This however actually yielded an error for me.
After some search, I found that one needs to generate the documentation
in vim via `helptags`. This needs to be done once. I added a respective
entry in my installer script, but also an entry in the known issues, in
case this fails for some reasons.