fix(.bashrc): pin python version only in $HOME
Apparently `uv python ping` creates a `.python-version`-file, where it
is executed. Ordinary when opening a shell, this is a not a problem, as
this happens ordinary in $HOME. But if one e.g. uses a shortcut to open
a shell via the Finder (MacOS) or the Explorer (Windows), this prevents
uv from creating these files at random places.