feat(.vimrc): replace ALE with coc.nvim
I kept on encountering issues with ALE. Therefore I looked for an
alternative. `coc.nvim` seems to be the most advanced alterantive, which
unfortunately comes with the need of nodejs, but otherwise offers e.g.
an own installer feature. It still uses other LSPs and connects to those
using its own plugins. In addition, it is only supported with vim>=9.
For earlier versions, I either require an alternative or will not use
any. Right now ubuntu 24 and debian 12 both come with `vim`==9 out of
the box in `apt`.