feat(.bashrc): activate local uv venv in screen
Currently I am using per default my uv_base venv, which is supposed to
handle globally installed python utilities. However when entering
certain uv projects and usinv coc.nvim, it does not automatically detect
the local venv but relies instead on the activated uv_base venv.

For this reason I created a detection if I am currently using a screen
session and if a venv exist in the respective base-directory of the
screen-session, it is always activated instead of base_uv.

This is not perfect, but aligns with my light-weight remote development
setup, to create screen sessions in projects.